remote: Drop "snapshot"/"namespace" concept
The main idea behind the snapshot concept was to use both "file name"
and "image directory" to identify an image stored in cache. A unique
image directory is used as a "namespace" that allows to keep multiple
sets of images, which are produced by one or more pre-dump operations.

This concept was introduced to enable the --remote option to be used
with one or more pre-dump followed by a dump operation.

However, the idea of pre-dump is to extract only a set of memory pages
and enable the memory tracking mechanism, which can be used to
implement a pre-copy migration algorithm.

After the first pre-dump, a copy of all extracted memory will be send
to the destination host, and any further dump or pre-dump operations
will send only the modified memory pages.

Therefore, in the context of live migration, which is the main purpose
of the --remote option, it does not make sense to keep a copy of
obsolete data (previous memory pages) on the destination side.

A better solution that can be used to replace the "snapshot" concept
would be to apply/merge the modifications on the restore side at the
time of receiving, and overwrite previous values. Thus, keep only a
singe representation of all image files with up-to-date content, and
any previous cache data should be overwritten.

This patch only drops the "snapshot" concept, the above described idea
will be introduced in subsequent changes.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>