remote: Use type to identify images
Currently, all images are stored in a single double linked list, and
every time we want to check if an image is in the list we have to go
through all items, and use strncmp() to compare the image name (path).

A better approach would be use a data structure similar to a hash
table. However, there is no need to compute a hash since every image
already has an interger value (in the range from 0 to CR_FD_MAX), known
as type, that is associated with it. Thus, we can use this value as an
index into an array of buckets, where every bucket is a list of images.

This idea is implemented in a following patch, which replaces the
list rimg_head with an array of lists, and the current changes serve as
preparation for that.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>