remote: Replace rimg list with array of lists
Currently, we have a single list where all images are being added, and
each time we want to check if an image is in the list, we have to go
through every item in the list. Instead, use an array of lists where
the image type is used as an index. The reason for this is because
there could be multiple images of the same time. This is more efficient
because in order to identify an image, we need to iterate only through
a list images that have the same type.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>