remote: Replace rimg list with array of lists
The current data structure used to store image files in memory is
linked list which has linear access time. In other words, all received
images are added into a list and every time we want to check if an
image is in the list, we have to go through every item in the list.

In this patch is introduced an array of lists that aims to improve the
access time by using the image type as an index. Thus, in order to
identify an image, CRIU iterates only through a limited set of cached
images with specified type.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>