remote: Drop "snapshot"/"namespace" concept
The main idea behind the snapshot concept (a.k.a namespace) was to use
both "file name" and "image directory path" to identify images stored
in cache.

This concept was introduced to enable the --remote option to be used
with one or more pre-dump followed by a dump operation.

A pre-dump operation extracts only the memory pages of a process tree.
After the first pre-dump, a copy of all extracted memory is being send
over to the destination host, and any further pre-dump or dump
operations will send only the modified memory pages.

However, in the context of live migration, it is not very useful to
store the old version of memory pages on the destination side.

An alternative solution is to merge the received modifications of
memory pages with previously received data, and keep only a up-to-date
content necessary for to restore the process tree.

For simplicity, in this patch is removed only the "snapshot" concept.
The idea described above is implemented in subsequent commits.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>