remote: Drop snapshot_id concept
In order to support iterative migration, the concept of snapshot_id
was introduced (only when remote migration is enabled). Each image
is tagged with one snapshot_id. The snapshot_id is the image directory
used for the (pre-dump or dump) operation that creates the image.

Images stored in memory (both in image-proxy and image-cache) are
identified by their name and snapshot_id. Snapshot IDs are ordered
so that we can find parent pagemaps that will be used when restoring
the process.

However, in every pre-dump iteration are transferred only memory pages
(pagemaps) and no other images.

At first, all memory pages copied over to the destination host and any
further pre-dump (or dump) action will send only modified memory pages.

Therefore, in this context, it is not necessary to preserve the old
copy of a modified memory page on the destination host. It can be simply
overwritten by the new content that will be used during restore.

To make code review easier this patch only removes the snapshot_id
concept. All other necessary changes to fully support iterative
remote migration are implemented in the following commits.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>