remote: Use image type as an identifier
Currently, when CRIU is running in remote mode all image files are
stored in a linked list. The access time with this data structure is
O(N), that is, CRIU has to iterate through all elements in the list
in order to find a received image.

An alternative approach is to use an array of lists where the image
type is used as an index.

In this patch has been introduced the image type as an identifier and
in a following patch will be introduced the above described data
structure.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>