remote: Use type of image as identifier
In remote migration all image files are stored in a linked list and the
access time is O(N). Thus, we have to iterate through all elements in
the list in order to find out if an image has been received.

An alternative approach is to use array of lists where the image type
is used as index.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>