remote: Drop 'open_mode' attribute
Currently, both local_image and remote_image protobuf entry definitions
have a required attribute 'open_mode', which is used to transfer the
flags specified by CRIU when opening an image.

The main purpose of this attribute is to enable image-cache/proxy to
distinguish between read and write operations, assuming that dump
or pre-dump could perform both operations on a remote image.

However, this assumption is not relevant because dump/pre-dump does not
require read access to images on resotre side, except for the case of
previous (parent) image directory, which is not relevant for the
--remote option.

For simplicity, lets assume that (pre-)dump operations should require
write-only access to content that is being stored by image-cache, and
restore will require read-only access to that content.

Hence, there is no need to provide a mechanism for reading cached
images via image-proxy.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>