remote: Drop local_image_entry protobuf msg
Both message types, 'remote' and 'local' image entry, are used to send
name and type of an image. The only difference is that the former has
an additional attribute - 'size'.

However, there is no benefit of having both message types and it would
be better to combine them by making the 'size' attribute optional.

Note that if a default value for an optional element is not specified,
a type-specific default value is used instead. For numeric types, the
default value is zero.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>