remote: Drop remote operation close_fd attribute
The close_fd attribute of the roperation structure is used to indicate
whether the file descriptor of the operation is a local or remote. This
is unnecessary since remote_sk is a available to functions in
img-remote.c file. Thus, in order to achieve the same result it is
sufficient to check if the file descriptor of the remote operation is
equal to remote_sk.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>