remote: Drop 'open_mode' protobuf attribute
image-cache and image-proxy use the 'open_mode' attribute of
local_image/remote_image protocol buffer definitions to indicate
if an image should be sent or received.

However, when --remote is used with dump (or pre-dump) we don't
read images stored on the destination host anymore (this was only
necessary in the case when --prev-images-dir was used).

Therefore, if we assume that (pre-)dump requires write-only access
and restore requires read-only access, the 'open_mode' attribute
becomes unnecessary.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>