arm: Removed SP from clobber list
Clobbering the stack pointer register is deprecated.  GCC expects
the value of the stack pointer after an asm statement to be the same
as it was before, so no asm can validly clobber the stack pointer in
the usual sense.  Adding the stack pointer to the clobber list has
traditionally had some undocumented and somewhat obscure side-effects.

https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=99063eeecbfb15fece3c72747a9dd14077c41afd
https://github.com/ARM-software/CMSIS_5/commit/a07b56505471fd789111bd7dba726461449cc261

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>