tls: enable encryption for protobuf images
This patch extends the dump and pre-dump functionality with the ability
to encrypt CRIU images using Authenticated Encryption with Associated
Data (Chacha20 cipher with the Poly1305 authenticator) in combination
with the existing public-key support for TLS.

When the '--tls' option is specified during checkpoint, CRIU will
attempt to load a public key from X.509 certificate. If the public key
is successfully loaded, CRIU generates a 256-bit random *token* that is
encrypted with the public key and saved in 'cipher.img'. This token is
then used to encrypt the content of all other CRIU images.

The 'criu restore' command is also extended with support for encrypted
CRIU images. When the `--tls` option is used together with restore
operation, CRIU will attempt to load an asymmetric private key and use
it to decrypt the token value from 'cipher.img'. The token value is
then used to decrypt all other CRIU images.

Signed-off-by: Radostin Stoyanov <radostin.stoyanov@eng.ox.ac.uk>