tls: enable protobuf image encryption
This patch extends the dump, pre-dump functionality with the ability to
encrypt CRIU images using Authenticated Encryption with Associated Data
(Chacha20 cipher with the Poly1305 authenticator) in combination with
the existing public-key support for TLS.

When the '--tls' option is used during checkpoint CRIU will will try to
load a public key (PK) from an X.509 certificate. If it successfully
loaded, CRIU will generate a 256-bit random token that is then encrypted
with PK and stored in 'cipher.img'. This token is then used to encrypt
the content of all images created by CRIU.

Signed-off-by: Radostin Stoyanov <radostin.stoyanov@eng.ox.ac.uk>