tls: enable encrypted image restore
This patch extends the `criu restore` command with support for the
`--tls` option to enable the restore operation using encrypted
images. When this option is used, CRIU will load a corresponding
asymmetric private key and decrypt the token value from 'cipher.img'.
The decrypted token value is then used to decrypt the protobuf entries
of all other CRIU images.

Signed-off-by: Radostin Stoyanov <radostin.stoyanov@eng.ox.ac.uk>