config: add --action-env cli option
In some use cases, action scripts might require user-specified
environment variables to be passed. For example, when creating a
checkpoint of a Pod, we would like to be able to pass the container
name and pod name to the action script.

To enable this functionality, this patch adds a new CLI option
`--action-env` that takes as input a string in the following format
"<NAME>=<VALUE>". This option can be used multiple times.

Example:

	criu dump --action-script /foo/bar/script.sh \
		--action-env "FOO1=BAR1" \
		--action-env "FOO2=BAR2" ...

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>