cr_system: enable data encryption
cr_system() and cr_system_userns() are used to run external executables
such as tar, ip, and iptables. These external tools are used to create
image files in 3rd party format (i.e., raw images). In order to encrypt
the output of these tools, and to decrypt their input, we introduce two
auxiliary functions that replace the corresponding input or output file
descriptor with a pipe, and perform encryption/decryption of the data.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>