Replace flake8 with ruff
ruff (https://github.com/astral-sh/ruff) is a Python linter
written in Rust that is designed to replace for Flake8.
It is an order of magnitude faster and actively maintained.

In addition to replacing flake8 with ruff, this patch also
creates separate makefile targets for ruff, shellcheck and
codespell, so that they can be tested independently.

The RUFF_FLAGS can be used to specify options such as '--fix'.
Example:
	make lint
	make ruff RUFF_FLAGS=--fix

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>