tls: verify the integrity of memory pages
The AES-XTS cipher does not provide integrity verification.
In this patch we add a verification mechanism based on the
HMAC-SHA-256 algorithm.

In order to support iterative checkpointing and memory deduplication
with encrypted memory, and to avoid storing HMAC for each memory page,
we compute XOR for of HMAC value for all memory pages and store this
value in cipher.img

The XOR computation also allows us to address the problem that memory
pages are read during restore in a different order then they are written
during checkpoint. In addition, to ensure that memory pages are restored
in correct order, we include the PID and VMA address associated with each
page in the HMAC computation.

The following example illustrates the HMAC value computation:

	H_n = HMAC(PID + VMA + MEMORY + KEY)
	hmac_value = H_1 ^ H_2 ^ ... ^ H_n

- PID: PID associated with the memory page
- VMA: virtual memory address associated with memory page
- KEY: secret key
- H_n: n-th memory page
- hmac_value: value stored in cipther.img during checkpoint, and used
              for integrity verification during restore

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>