images/inventory: add plugins field
This patch extends the inventory image with a `plugins` field
that contains an array of names. This field indicates which
plugins were used during checkpoint, for example, to save GPU
state. In particular, the CUDA and AMDGPU plugins are added
to this field only when the checkpoint contains GPU state.
This allows to migrate a process that does not use GPU
from a GPU-enabled system to CPU-only environment.

During restore, this field is used to disable unnecessary
plugins and show appropriate error messages if required
CRIU plugin are missing.

To preserve backwards compatibility, we use an `optional`
`plugins_entry` that allows us to distinguish between an
empty and unset `plugins` field.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>