cuda: enable checkpoint support for paused tasks
If a CUDA process is already in a "locked" or "checkpointed" state
during criu dump, the CUDA plugin fails with an error because it
attempts an unnecessary "lock" action using the cuda-checkpoint tool.

This patch extends the CUDA plugin to handle such case by first checking
the original state of the CUDA process and skipping unnecessary "lock"
and "checkpoint" actions if the process was already locked or
checkpointed before CRIU was invoked.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>