tls: enable support for encryption of memory pages
We use the AES-XTS block cipher to encrypt memory pages as it is
designed to encrypt blocks of data with fixed-size (e.g. memory pages),
allows the use of hardware acceleration available in modern CPUs, and
uses a single initialization vector (IV), instead of per-page nonce,
to ensure that encrypting the same plaintext with the same key results
in different ciphertexts.

In particular, XTS uses two 256-bits AES keys. One key is used to
perform block encryption, and the other is used to encrypt a so-called
"tweak value". The encrypted tweak value is further modified (with a
Galois polynomial function) and XOR-ed with both the plaintext and
ciphertext of each block. This method ensures that encrypting multiple
blocks with identical data will produce different ciphertext.

Since CRIU restores memory pages in the restorer context, this PIE code
cannot be linked with libraries such as GnuTLS to perform decryption.
Instead, we introduce a helper process to decrypt memory pages data.
The restorer context communicates with this helper process using PIPEs.
It sends the function arguments be used by preadv() and receives back
its return value. The decrypted data is transferred to the target
address space with process_vm_writev.

Suggested-by: Daiki Ueno <dueno@redhat.com>
Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>