cr-service: refactor early rpc config parsing
When an additional configuration file is specified via RPC, this file is
parsed twice: first at an early stage to load options that are important
for debugging, such as --log-file, --work-dir, and --images-dir, and
again after all other methods for specifying config options have been
evaluated, for example, to overwrite options specified by the container
runtime (e.g., --tcp-established). However, this leads to silently
duplicating values of repeatable options (e.g., --action-script).
To address this problem, we limit the number of options that are
parsed at the early stage.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>